Move Mongo connection code into main
Although there's no effective difference to the code (they're both
compiled into the same package), this achieves a few things:

- Groups environment and global variables
- Makes it clear where the setup/teardown code for the database
  connection is
