+++ handlers.go
@@ -4,33 +4,14 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"sync"
 	"time"
 
 	"labix.org/v2/mgo"
 	"labix.org/v2/mgo/bson"
 )
 
-var (
-	mgoSession      *mgo.Session
-	mgoSessionOnce  sync.Once
-	mgoDatabaseName = getenvDefault("LINK_TRACKER_MONGO_DB", "external_link_tracker")
-	mgoURL          = getenvDefault("LINK_TRACKER_MONGO_URL", "localhost")
-)
-
-// Store function in a variable so it can be overridden in the tests.
+// store function in a variable so it can be overridden in the tests.
 var now = time.Now
-
-func getMgoSession() *mgo.Session {
-	mgoSessionOnce.Do(func() {
-		var err error
-		mgoSession, err = mgo.Dial(mgoURL)
-		if err != nil {
-			panic(err) // no, not really
-		}
-	})
-	return mgoSession.Clone()
-}
 
 type ExternalLink struct {
 	ExternalURL string `bson:"external_url"`
